use objected approach to AudioNotifier
