+++ src/ui/Controller.h
@@ -7,6 +7,7 @@
 #include "KeysWindows.h"
 #include "ConfigModel.h"
 #include "BackupsModel.h"
+#include "AudioNotifier.h"
 #include "../parse/Parse.h"
 
 // Handler of hotkeys presss
@@ -68,7 +69,6 @@ struct ERCharInfoResult {
     std::vector<fssm::parse::er::ERCharacterInfo> characters;
 };
 
-class AudioNotifier;
 
 // Controller wrapping backend logic allowing UI to access data it needs
 class Controller: public QObject {
@@ -116,7 +116,7 @@ private slots:
     void onBackupLoad(bool success);
 
 private:
-    std::unique_ptr<AudioNotifier> m_audioNotifier;
+    AudioNotifier m_audioNotifier;
     QString m_currentSaveId = "";
     ConfigModel* m_configModel;
     BackupsModel* m_backupsModel;
